evix: centralize distributed scheduler coordination
